# backend/docs/api/openapi.pretty.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Niyat Ride API",
        "description": "API Documentation for Niyat Ride",
        "contact": {
            "name": "Niyat Support",
            "email": "backendteam@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/admin/dispatchers/{id}/status": {
            "patch": {
                "tags": [
                    "Dispatcher User Management"
                ],
                "summary": "Update dispatcher user status",
                "operationId": "updateDispatcherStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StatusUpdateRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/DispatcherUserResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/customers/{id}/status": {
            "patch": {
                "tags": [
                    "Customer User Management"
                ],
                "summary": "Update customer user status",
                "operationId": "updateCustomerStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StatusUpdateRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerUserResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/dispatchers": {
            "get": {
                "tags": [
                    "Dispatcher User Management"
                ],
                "summary": "Get all dispatcher users with filtering and pagination",
                "operationId": "getAllDispatchers",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 20
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "createdAt"
                        }
                    },
                    {
                        "name": "sortDirection",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "PENDING",
                                "ACTIVE",
                                "SUSPENDED",
                                "DEACTIVATED",
                                "DELETED"
                            ]
                        }
                    },
                    {
                        "name": "createdAtFrom",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "createdAtTo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/dispatchers/{id}": {
            "get": {
                "tags": [
                    "Dispatcher User Management"
                ],
                "summary": "Get dispatcher user by ID",
                "operationId": "getDispatcherById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/DispatcherUserResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/customers": {
            "get": {
                "tags": [
                    "Customer User Management"
                ],
                "summary": "Get all customer users with filtering and pagination",
                "operationId": "getAllCustomers",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 20
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "createdAt"
                        }
                    },
                    {
                        "name": "sortDirection",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "PENDING",
                                "ACTIVE",
                                "SUSPENDED",
                                "DEACTIVATED",
                                "DELETED"
                            ]
                        }
                    },
                    {
                        "name": "createdAtFrom",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "createdAtTo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/customers/{id}": {
            "get": {
                "tags": [
                    "Customer User Management"
                ],
                "summary": "Get customer user by ID",
                "operationId": "getCustomerById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerUserResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StatusUpdateRequestDTO": {
                "required": [
                    "status"
                ],
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "PENDING",
                            "ACTIVE",
                            "SUSPENDED",
                            "DEACTIVATED",
                            "DELETED"
                        ]
                    }
                }
            },
...
            "DispatcherUserResponseDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "isVerified": {
                        "type": "boolean"
                    },
                    "verifiedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "DISPATCHER",
                            "DRIVER",
                            "Admin",
                            "CUSTOMER"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "PENDING",
                            "ACTIVE",
                            "SUSPENDED",
                            "DEACTIVATED",
                            "DELETED"
                        ]
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "assignedRegion": {
                        "type": "string"
                    }
                }
            },
            "CustomerUserResponseDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "isVerified": {
                        "type": "boolean"
                    },
                    "verifiedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "DISPATCHER",
                            "DRIVER",
                            "Admin",
                            "CUSTOMER"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "PENDING",
                            "ACTIVE",
                            "SUSPENDED",
                            "DEACTIVATED",
                            "DELETED"
                        ]
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
        }
    }
}